set/module in set methods, moved some impl around
